// packages/mpl-core-kit-lib/src/generated/types/baseLinkedAppData.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import { type Codec, combineCodec, type Decoder, type Encoder, getStructDecoder, getStructEncoder } from '@solana/kit'

import {
  type BasePluginAuthority,
  type BasePluginAuthorityArgs,
  type ExternalPluginAdapterSchema,
  type ExternalPluginAdapterSchemaArgs,
  getBasePluginAuthorityDecoder,
  getBasePluginAuthorityEncoder,
  getExternalPluginAdapterSchemaDecoder,
  getExternalPluginAdapterSchemaEncoder,
} from './index.ts'

export type BaseLinkedAppData = {
  dataAuthority: BasePluginAuthority
  schema: ExternalPluginAdapterSchema
}

export type BaseLinkedAppDataArgs = {
  dataAuthority: BasePluginAuthorityArgs
  schema: ExternalPluginAdapterSchemaArgs
}

export function getBaseLinkedAppDataCodec(): Codec<BaseLinkedAppDataArgs, BaseLinkedAppData> {
  return combineCodec(getBaseLinkedAppDataEncoder(), getBaseLinkedAppDataDecoder())
}

export function getBaseLinkedAppDataDecoder(): Decoder<BaseLinkedAppData> {
  return getStructDecoder([
    ['dataAuthority', getBasePluginAuthorityDecoder()],
    ['schema', getExternalPluginAdapterSchemaDecoder()],
  ])
}

export function getBaseLinkedAppDataEncoder(): Encoder<BaseLinkedAppDataArgs> {
  return getStructEncoder([
    ['dataAuthority', getBasePluginAuthorityEncoder()],
    ['schema', getExternalPluginAdapterSchemaEncoder()],
  ])
}